Auth and notification system

+++ app/Http/Controllers/Admin/ProjectController.php
@@ -9,20 +9,35 @@ use App\Http\Requests\UpdateProjectRequest;
 use App\Models\Client;
 use App\Models\Project;
 use App\Models\User;
+use App\Notifications\ProjectAssigned;
+use App\Notifications\ProjectUnAssigned;
+use App\Notifications\TaskAssigned;
+use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Notification;
+use Illuminate\Validation\Rules\NotIn;
+use Symfony\Component\VarDumper\VarDumper;
+use App\Services\NotificationService;
 
 class ProjectController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
+    // this method does not need protection - because we assume that we have just ( admin - user ) roles and we handle this using if-else
     public function index()
     {
-        $projects = Project::all();
+        if(auth()->user()->hasRole('admin')){
+            $projects = Project::all();
+        } else {
+            $user = Auth::user();
+            $projects = $user->Projects()->get();
+        }
         return view('admin.projects.index', [
             'projects' => $projects,
             'page' => 'projects List'
-        ]);   
+        ]);
     }
 
     /**
@@ -30,6 +45,8 @@ class ProjectController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Project::class);
+
         $users = User::all();
         $clients = Client::all();
         return view('admin.projects.create', [
@@ -44,10 +61,21 @@ class ProjectController extends Controller
      */
     public function store(ProjectRequest $request)
     {
-        Project::create($request->validated());
+        // $this->authorize('restore');
+
+        $assignedUsers = $request->input('assigned_users');
+        if( sizeof($assignedUsers)>0){
+            $project = Project::create($request->validated());
+
+            foreach ($assignedUsers as $assignedUser) {
+                $project->users()->attach($assignedUser);
+            }
+            
+        } else {
+            return redirect()->back()->with('message', 'please do not leave the project without users');
+        }
         
         return redirect()->route('admin.projects.index')->with('message', 'the project has been created sucessfully');;
-    
     }
 
     /**
@@ -55,6 +83,8 @@ class ProjectController extends Controller
      */
     public function show(Project $project)
     {
+        $this->authorize('view', $project);
+
         $project->with('user', 'client');
         return view('admin.projects.show', [
             'project' => $project,
@@ -67,6 +97,8 @@ class ProjectController extends Controller
      */
     public function edit(Project $project)
     {
+        $this->authorize('update', $project);
+
         $users = User::all();
         $clients = Client::all();
         return view('admin.projects.edit', [
@@ -82,6 +114,8 @@ class ProjectController extends Controller
      */
     public function update(UpdateProjectRequest $request, Project $project)
     {
+        $this->authorize('update', $project);   
+        
         $project->update([
             'title'       => $request->validated('title'),
             'description' => $request->validated('description'),
@@ -90,6 +124,7 @@ class ProjectController extends Controller
             'client_id'   => $request->validated('client_id'),
         ]);
         
+
         $assignedUsers = $request->input('assigned_users');
         if( sizeof($assignedUsers)>0){
             $project->users()->detach();
@@ -115,10 +150,13 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
+        $this->authorize('delete', $project);
+
         $project->delete();
         return redirect()->route('admin.projects.index')->with('message','the project has been deleted successfully');
     }
 
+    // protect this method using meddleware
     public function assignCreate(Project $project)
     {
         $users = User::all();
@@ -130,15 +168,47 @@ class ProjectController extends Controller
         ]);
     }
 
+    // protect this method using meddleware
     public function assignStore(AssignUserStoreRequest $request, Project $project)
     {
-        $assignedUsers = $request->input('assigned_users');
-        if( sizeof($assignedUsers)>0){
-            $project->users()->detach();
-            foreach ($assignedUsers as $assignedUser) {
-                $project->users()->attach($assignedUser);
-            }
-        }
-        return redirect()->route('admin.projects.index')->with('message', 'the project has been updated sucessfully with new users');;
+        // the users those are been passed to here form the page  
+        $assignedUsers = $request->input('assigned_users'); //associative array 
+
+        $sendNotification = new NotificationService($project, $assignedUsers);
+        $sendNotification->SendNotificationMessages();
+        // //get the ids of the users those are already in the project 
+        // $projectUsers = $project->users()->pluck('users.id');
+
+        // if($assignedUsers != null && sizeof($assignedUsers)>0){    
+        //     // get all the users 
+        //     $users = User::all();
+        //     $usersIds= $users->pluck('id');
+        //     foreach ($usersIds as $usersId){
+        //         if($projectUsers->contains($usersId) && !in_array($usersId,$assignedUsers)){
+        //             // delete notification
+        //             $user = User::findOrFail($usersId);
+        //             $user->notify(new ProjectUnAssigned($project));
+        //         } else if (!$projectUsers->contains($usersId) && in_array($usersId,$assignedUsers)){
+        //             // assing notification
+        //             $user = User::findOrFail($usersId);
+        //             $user->notify(new ProjectAssigned($project));
+        //         } else {
+        //             // nothing
+        //         }
+        //     }
+
+        //     $project->users()->detach();
+        //     foreach ($assignedUsers as $assignedUser) {
+        //         $project->users()->attach($assignedUser);
+        //     }
+        // } else {
+        //     //special case when there is just one user in the project and you remove it - then the assignedUsers will be null
+        //     $users = $project->users()->get();
+        //     Notification::send($users, new ProjectUnAssigned($project));
+        //     // if the admin uncheck all users 
+        //     $project->users()->detach();
+        // }
+
+        return redirect()->route('admin.projects.index')->with('message', 'the project has been updated sucessfully');
     }
 }
